+++ backend/src/main/java/br/ifsul/finwise/repository/ProfessorRepositorio.java
@@ -4,12 +4,13 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
+
+import br.ifsul.finwise.model.ProfessorModelo;
 import br.ifsul.finwise.model.UsuarioModelo;
 import java.util.List;
-import java.util.Optional;
 
 @Repository
-public interface ProfessorRepositorio extends JpaRepository<ProfessorModelo, Long> {
+public interface ProfessorRepositorio extends JpaRepository<ProfessorModelo, Integer> {
 
     // Busca usuários contendo parte do nome (ignora maiúsculas/minúsculas)
     List<UsuarioModelo> findByNameContainingIgnoreCase(String name);
@@ -22,7 +23,7 @@ public interface ProfessorRepositorio extends JpaRepository<ProfessorModelo, Lon
 
     // Buscar status do professor por id
     @Query("SELECT p FROM ProfessorModelo WHERE p.id = id")
-    List<ProfessorModelo> findStatusById(@Param("id") Long id);
+    List<ProfessorModelo> findStatusById(@Param("id") Integer id);
 
     // Lista usuários em ordem alfabética
     @Query("SELECT u FROM UsuarioModelo u ORDER BY u.name ASC")
@@ -34,31 +35,31 @@ public interface ProfessorRepositorio extends JpaRepository<ProfessorModelo, Lon
 
     // Busca usuários por múltiplos IDs
     @Query("SELECT u FROM UsuarioModelo u WHERE u.id IN :ids")
-    List<UsuarioModelo> findByIds(@Param("ids") List<Long> ids);
+    List<UsuarioModelo> findByIds(@Param("ids") List<Integer> ids);
 
      // Busca foto de perfil do usuário por ID
     @Query("SELECT u.fotoPerfil FROM ProfessorModelo u WHERE u.id = :id")
-    byte[] findfotoPerfilByUserId(@Param("id") Long id);
+    byte[] findfotoPerfilByUserId(@Param("id") Integer id);
 
     // Busca biografia do usuário por ID
     @Query("SELECT u.biografia FROM ProfessorModelo u WHERE u.id = :id")
-    String findbiografiaByUserId(@Param("id") Long id);
+    String findbiografiaByUserId(@Param("id") Integer id);
 
     // Atualiza nome do usuário
     @Query("UPDATE UsuarioModelo u SET u.name = :newName WHERE u.id = :id")
-    int updateUserNameById(@Param("id") Long id, @Param("newName") String newName);
+    int updateUserNameById(@Param("id") Integer id, @Param("newName") String newName);
 
     // Atualiza senha do usuário
     @Query("UPDATE UsuarioModelo u SET u.password = :newPassword WHERE u.id = :id")
-    int updateUserPasswordById(@Param("id") Long id, @Param("newPassword") String newPassword);
+    int updateUserPasswordById(@Param("id") Integer id, @Param("newPassword") String newPassword);
 
     // Atualiza foto de perfil do usuário
     @Query("UPDATE ProfessorModelo u SET u.fotoPerfil = :fotoPerfil WHERE u.id = :id")
-    int updateUserfotoPerfilById(@Param("id") Long id, @Param("fotoPerfil") byte[] fotoPerfil);
+    int updateUserfotoPerfilById(@Param("id") Integer id, @Param("fotoPerfil") byte[] fotoPerfil);
 
     // Atualiza biografia do usuário
     @Query("UPDATE ProfessorModelo u SET u.biografia = :biografia WHERE u.id = :id")
-    int updateUserbiografiaById(@Param("id") Long id, @Param("biografia") String biografia);
+    int updateUserbiografiaById(@Param("id") Integer id, @Param("biografia") String biografia);
 
     // Exclui usuários pelo nome
     @Query("DELETE FROM UsuarioModelo u WHERE u.name = :name")
@@ -66,9 +67,9 @@ public interface ProfessorRepositorio extends JpaRepository<ProfessorModelo, Lon
 
     // Remove foto de perfil do usuário
     @Query("UPDATE ProfessorModelo u SET u.fotoPerfil = null WHERE u.id = :id")
-    int removeUserfotoPerfilById(@Param("id") Long id);
+    int removeUserfotoPerfilById(@Param("id") Integer id);
 
     // Remove descrição do usuário
     @Query("UPDATE ProfessorModelo u SET u.biografia = null WHERE u.id = :id")
-    int removeUserbiografiaById(@Param("id") Long id);
+    int removeUserbiografiaById(@Param("id") Integer id);
 }
